+++ tests/integ/scala/test_dataframe_writer_suite.py
@@ -8,7 +8,11 @@ import pytest
 
 from snowflake.snowpark import Row
 from snowflake.snowpark._internal.utils import parse_table_name
-from snowflake.snowpark.exceptions import SnowparkSQLException
+from snowflake.snowpark.exceptions import (
+    SnowparkDataframeWriterException,
+    SnowparkSQLException,
+)
+from snowflake.snowpark.functions import col, lit, object_construct
 from snowflake.snowpark.types import (
     DoubleType,
     IntegerType,
@@ -16,7 +20,222 @@ from snowflake.snowpark.types import (
     StructField,
     StructType,
 )
-from tests.utils import TestFiles, Utils
+from tests.utils import TestData, TestFiles, Utils
+
+
+def test_write_to_csv(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.csv(f"@{target_stage_name}/test1.csv")
+    res = (
+        session.read.schema(df.schema)
+        .csv(f"@{target_stage_name}/test1.csv")
+        .orderBy("NUM")
+        .collect()
+    )
+    assert res == [Row(1, True, "a"), Row(2, False, "b")]
+
+
+def test_write_to_csv_ignore_overwrite_when_non_bool_non_str(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.csv(f"@{target_stage_name}/test1.csv", overwrite={})
+    res = (
+        session.read.schema(df.schema)
+        .csv(f"@{target_stage_name}/test1.csv")
+        .orderBy("NUM")
+        .collect()
+    )
+    assert res == [Row(1, True, "a"), Row(2, False, "b")]
+
+
+def test_write_to_csv_normal_options_non_aliases(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.option("FIELD_DELIMITER", ";").option("RECORD_DELIMITER", "\n").option(
+        "FIELD_OPTIONALLY_ENCLOSED_BY", "'"
+    ).option("NULL_IF", "b").option("DATE_FORMAT", "YYYY-MM-DD").csv(
+        f"@{target_stage_name}/test_csv_normal_options_non_aliases.csv", overwrite={}
+    )
+    res = (
+        session.read.schema(df.schema)
+        .option("FIELD_DELIMITER", ";")
+        .option("RECORD_DELIMITER", "\n")
+        .option("FIELD_OPTIONALLY_ENCLOSED_BY", "'")
+        .option("NULL_IF", "b")
+        .csv(f"@{target_stage_name}/test_csv_normal_options_non_aliases.csv")
+        .orderBy("NUM")
+        .collect()
+    )
+    assert res == [Row(1, True, "a"), Row(2, False, None)]
+
+
+def test_write_to_csv_with_aliases(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.option("DELIMITER", "x").option("SEP", ";").option("LINESEP", "\n").option(
+        "QUOTE", "'"
+    ).option("NULLVALUE", "b").option("DATEFORMAT", "YYYY-MM-DD").csv(
+        f"@{target_stage_name}/test1.csv", mode="overwrite"
+    )
+    res = (
+        session.read.schema(df.schema)
+        .option("SEP", ";")
+        .option("LINESEP", "\n")
+        .option("QUOTE", "'")
+        .option("NULLVALUE", "b")
+        .option("DATEFORMAT", "YYYY-MM-DD")
+        .csv(f"@{target_stage_name}/test1.csv")
+        .orderBy("NUM")
+        .collect()
+    )
+    assert res == [Row(1, True, "a"), Row(2, False, None)]
+
+
+def test_write_to_csv_with_options(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.option("path", f"@{target_stage_name}/test1.csv").option(
+        "overwrite", True
+    ).option("single", True).option("header", True).option(
+        "MAX_FILE_SIZE", 5000
+    ).option(
+        "DETAILED_OUTPUT", True
+    ).csv(
+        f"@{target_stage_name}/test2.csv"
+    )
+    df.write.option("path", f"@{target_stage_name}/test1.csv").option(
+        "overwrite", True
+    ).option("overwrite", False).option("overwrite", "True").option(
+        "overwrite", "False"
+    ).option(
+        "single", False
+    ).option(
+        "header", True
+    ).csv(
+        f"@{target_stage_name}/test1.csv",
+        format="csv",
+        mode="overwrite",
+        partitionBy=col(df.schema.fields[0].name),
+        block=True,
+        sep=",",
+    )
+    res = (
+        session.read.schema(df.schema)
+        .option("header", True)
+        .csv(f"@{target_stage_name}/test1.csv")
+        .orderBy("NUM")
+        .collect()
+    )
+    assert res == [Row(1, True, "a"), Row(2, False, "b")]
+
+
+def test_write_to_csv_using_save(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.option("path", f"@{target_stage_name}/test1.csv").save(
+        format="csv",
+        mode="overwrite",
+        partitionBy=df.schema.fields[0].name,
+        block=True,
+        sep=",",
+        overwrite=True,
+    )
+    res = (
+        session.read.schema(df.schema)
+        .csv(f"@{target_stage_name}/test1.csv")
+        .orderBy("NUM")
+        .collect()
+    )
+    assert res == [Row(NUM=1, BOOL=True, STR="a"), Row(NUM=2, BOOL=False, STR="b")]
+
+
+def test_write_to_csv_using_save_without_fileformat_should_fail(session):
+    with pytest.raises(
+        SnowparkDataframeWriterException,
+        match="DataFrameWriter.format()",
+    ):
+        target_stage_name = Utils.random_stage_name()
+        Utils.create_stage(session, target_stage_name)
+        df = TestData.test_data1(session)
+        df.write.option("path", f"@{target_stage_name}/test1.csv").save(
+            mode="overwrite",
+            partitionBy=df.schema.fields[0].name,
+            block=True,
+            sep=",",
+            overwrite=True,
+        )
+
+
+def test_write_to_csv_using_save_options(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.options(
+        {
+            "format": "csv",
+            "mode": "overwrite",
+            "partitionBy": df.schema.fields[0].name,
+            "sep": ",",
+        }
+    ).save(f"@{target_stage_name}/test1.csv")
+    res = (
+        session.read.schema(df.schema)
+        .csv(f"@{target_stage_name}/test1.csv")
+        .orderBy("NUM")
+        .collect()
+    )
+    assert res == [Row(1, True, "a"), Row(2, False, "b")]
+
+
+def test_write_to_format_csv(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.format("csv").csv(f"@{target_stage_name}/test1.csv")
+    res = (
+        session.read.schema(df.schema)
+        .csv(f"@{target_stage_name}/test1.csv")
+        .orderBy("NUM")
+        .collect()
+    )
+    assert res == [Row(1, True, "a"), Row(2, False, "b")]
+
+
+def test_write_to_json(session):
+    target_stage_name = Utils.random_stage_name()
+    target_stage_name = "stage1"
+    Utils.create_stage(session, target_stage_name, is_temporary=False)
+    df = TestData.test_data1(session)
+    df.select(
+        object_construct(lit("NUM"), "NUM", lit("BOOL"), "BOOL", lit("STR"), "STR")
+    ).write.json(f"@{target_stage_name}/test1.json")
+    res = session.read.json(f"@{target_stage_name}/test1.json")
+    res = (
+        session.read.json(f"@{target_stage_name}/test1.json")
+        .orderBy(col("$1")["NUM"])
+        .collect()
+    )
+    assert res == [
+        Row('{\n  "BOOL": true,\n  "NUM": 1,\n  "STR": "a"\n}'),
+        Row('{\n  "BOOL": false,\n  "NUM": 2,\n  "STR": "b"\n}'),
+    ]
+
+
+def test_write_to_parquet(session):
+    target_stage_name = Utils.random_stage_name()
+    Utils.create_stage(session, target_stage_name)
+    df = TestData.test_data1(session)
+    df.write.parquet(f"@{target_stage_name}/test1.parquet")
+    res = session.read.parquet(f"@{target_stage_name}/test1.parquet")
+    res = res.orderBy(res.columns[0]).collect()
+    assert res == [Row(1, True, "a"), Row(2, False, "b")]
 
 
 def test_write_with_target_column_name_order(session):