+++ tests/unit/scala/test_error_message.py
@@ -9,6 +9,7 @@ from snowflake.snowpark.exceptions import (
     SnowparkCreateViewException,
     SnowparkDataframeException,
     SnowparkDataframeReaderException,
+    SnowparkDataframeWriterException,
     SnowparkFetchDataException,
     SnowparkInvalidObjectNameException,
     SnowparkJoinException,
@@ -187,6 +188,26 @@ def test_merge_table_action_already_specified():
     assert type(ex) == SnowparkTableException
     assert ex.error_code == "1115"
     assert ex.message == f"{action} has been specified for {clause} to merge table"
+
+
+def test_df_must_provide_filetype_for_reading_file():
+    ex = SnowparkClientExceptionMessages.DF_MUST_PROVIDE_FILETYPE_FOR_READING_FILE()
+    assert type(ex) == SnowparkDataframeReaderException
+    assert ex.error_code == "1116"
+    assert (
+        ex.message
+        == "You must call DataFrameReader.format() or specify the file type for the file using the option() or load() with 'TYPE' or 'FORMAT'"
+    )
+
+
+def test_df_must_provide_filetype_for_writing_file():
+    ex = SnowparkClientExceptionMessages.DF_MUST_PROVIDE_FILETYPE_FOR_WRITING_FILE()
+    assert type(ex) == SnowparkDataframeWriterException
+    assert ex.error_code == "1117"
+    assert (
+        ex.message
+        == "You must call DataFrameWriter.format() or specify the file type for the file using the option() with 'TYPE' or 'FORMAT'"
+    )
 
 
 def test_plan_analyzer_invalid_identifier():