+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### New Features
 
+- Added support for aliases for commonly used options in DataFrame reader/writer.
+- Added support for `format(...).load()` and `format(...).save(...)`patterns.
+- Added support for `df.write.csv("path/to/stage")`,`df.write.json("path/to/stage")`,`df.write.parquet("path/to/stage")` patterns to allow a similar experience to dataframe readers.
 - Added support for new functions in `snowflake.snowpark.functions`:
   - `array_sort`
   - `sort_array`